+++ spec/spec/element.js
@@ -215,7 +215,7 @@ describe('Element', function() {
       rect.transform({ scaleY: 3 })
       expect(window.matrixStringToArray(rect.node.getAttribute('transform'))).toEqual([1,0,0,3,0,-100])
     })
-    
+
     it('performs an absolute scale by default', function() {
       rect.transform({ scale: 3 }).transform({ scale: 0.5 })
       expect(window.matrixStringToArray(rect.node.getAttribute('transform'))).toEqual([0.5,0,0,0.5,25,25])
@@ -460,7 +460,7 @@ describe('Element', function() {
       }))
     })
 
-    it('ungroups everything to the doc root when called on SVG.Doc / does not ungroup defs', function() {
+    it('ungroups everything to the doc root when called on SVG.Doc / does not ungroup defs/parser', function() {
       draw.ungroup()
       expect(rect1.parent()).toBe(draw)
       expect(rect2.parent()).toBe(draw)
@@ -476,10 +476,10 @@ describe('Element', function() {
         a:0.5, b:0, c:0, d:0.5, e:45, f:45
       }))
 
-      expect(draw.children().length).toBe(3) // 2 * rect + defs
+      expect(draw.children().length).toBe(3+parserInDoc) // 2 * rect + defs
     })
   })
-  
+
   describe('flatten()', function() {
     it('redirects the call to ungroup()', function() {
       spyOn(draw, 'ungroup')
@@ -819,8 +819,15 @@ describe('Element', function() {
 
         // Test for different browsers namely Firefox and Chrome
         expect(
+            // IE
             toBeTested === '<svg xmlns:svgjs="http://svgjs.com/svgjs" xmlns:xlink="http://www.w3.org/1999/xlink" version="1.1" xmlns="http://www.w3.org/2000/svg" height="100" width="100" id="' + draw.id() + '"><rect height="100" width="100"></rect><circle fill="#ff0066" cy="50" cx="50" r="50"></circle></svg>'
-         || toBeTested === '<svg id="' + draw.id() + '" width="100" height="100" xmlns="http://www.w3.org/2000/svg" version="1.1" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:svgjs="http://svgjs.com/svgjs"><rect width="100" height="100"></rect><circle r="50" cx="50" cy="50" fill="#ff0066"></circle></svg>').toBeTruthy()
+
+            // Firefox
+         || toBeTested === '<svg id="' + draw.id() + '" width="100" height="100" xmlns="http://www.w3.org/2000/svg" version="1.1" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:svgjs="http://svgjs.com/svgjs"><rect width="100" height="100"></rect><circle r="50" cx="50" cy="50" fill="#ff0066"></circle></svg>'
+
+            // svgdom
+         || toBeTested === '<svg id="' + draw.id() + '" xmlns="http://www.w3.org/2000/svg" version="1.1" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:svgjs="http://svgjs.com/svgjs" width="100" height="100"><svg id="SvgjsSvg1002" width="2" height="0" style="overflow: hidden; top: -100%; left: -100%; position: absolute; opacity: 0"><polyline id="SvgjsPolyline1003" points="10,10 20,10 30,10"></polyline><path id="SvgjsPath1004" d="M80 80A45 45 0 0 0 125 125L125 80Z "></path></svg><rect width="100" height="100"></rect><circle r="50" cx="50" cy="50" fill="#ff0066"></circle></svg>'
+        ).toBeTruthy()
 
       })
       it('returns partial raw svg when called on a sub group', function() {
@@ -832,7 +839,9 @@ describe('Element', function() {
 
         expect(
             toBeTested === '<g><rect width="100" height="100"></rect><circle r="50" cx="50" cy="50" fill="#ff0066"></circle></g>'
-         || toBeTested === '<g><rect height="100" width="100"></rect><circle fill="#ff0066" cy="50" cx="50" r="50"></circle></g>').toBeTruthy()
+         || toBeTested === '<g><rect height="100" width="100"></rect><circle fill="#ff0066" cy="50" cx="50" r="50"></circle></g>'
+         || toBeTested === '<g xmlns="http://www.w3.org/2000/svg"><rect width="100" height="100"></rect><circle r="50" cx="50" cy="50" fill="#ff0066"></circle></g>'
+        ).toBeTruthy()
       })
       it('returns a single element when called on an element', function() {
         var group = draw.group().id(null)
@@ -842,31 +851,37 @@ describe('Element', function() {
 
         expect(
             toBeTested === '<circle r="50" cx="50" cy="50" fill="#ff0066"></circle>'
-         || toBeTested === '<circle fill="#ff0066" cy="50" cx="50" r="50"></circle>').toBeTruthy()
+         || toBeTested === '<circle fill="#ff0066" cy="50" cx="50" r="50"></circle>'
+         || toBeTested === '<circle xmlns="http://www.w3.org/2000/svg" r="50" cx="50" cy="50" fill="#ff0066"></circle>'
+       ).toBeTruthy()
       })
     })
     describe('with raw svg given', function() {
       it('imports a full svg document', function() {
         draw.svg('<svg id="SvgjsSvg1000" xmlns="http://www.w3.org/2000/svg" version="1.1" xmlns:xlink="http://www.w3.org/1999/xlink" width="100" height="100" viewBox="0 0 50 50"><rect id="SvgjsRect1183" width="100" height="100"></rect><circle id="SvgjsCircle1184" r="50" cx="25" cy="25" fill="#ff0066"></circle></svg>')
-        expect(draw.get(0).type).toBe('svg')
-        expect(draw.get(0).children().length).toBe(2)
-        expect(draw.get(0).get(0).type).toBe('rect')
-        expect(draw.get(0).get(1).type).toBe('circle')
-        expect(draw.get(0).get(1).attr('fill')).toBe('#ff0066')
+
+        expect(draw.get(0+parserInDoc).type).toBe('svg')
+        expect(draw.get(0+parserInDoc).children().length).toBe(2)
+        expect(draw.get(0+parserInDoc).get(0).type).toBe('rect')
+        expect(draw.get(0+parserInDoc).get(1).type).toBe('circle')
+        expect(draw.get(0+parserInDoc).get(1).attr('fill')).toBe('#ff0066')
       })
       it('imports partial svg content', function() {
         draw.svg('<g id="SvgjsG1185"><rect id="SvgjsRect1186" width="100" height="100"></rect><circle id="SvgjsCircle1187" r="50" cx="25" cy="25" fill="#ff0066"></circle></g>')
-        expect(draw.get(0).type).toBe('g')
-        expect(draw.get(0).get(0).type).toBe('rect')
-        expect(draw.get(0).get(1).type).toBe('circle')
-        expect(draw.get(0).get(1).attr('fill')).toBe('#ff0066')
+        expect(draw.get(0+parserInDoc).type).toBe('g')
+        expect(draw.get(0+parserInDoc).get(0).type).toBe('rect')
+        expect(draw.get(0+parserInDoc).get(1).type).toBe('circle')
+        expect(draw.get(0+parserInDoc).get(1).attr('fill')).toBe('#ff0066')
       })
       it('does not import on single elements, even with an argument it acts as a getter', function() {
         var rect   = draw.rect(100,100).id(null)
           , result = rect.svg('<circle r="300"></rect>')
+
         expect(
             result === '<rect width="100" height="100"></rect>'
-         || result === '<rect height="100" width="100"></rect>').toBeTruthy()
+         || result === '<rect height="100" width="100"></rect>'
+         || result === '<rect xmlns="http://www.w3.org/2000/svg" width="100" height="100"></rect>'
+        ).toBeTruthy()
       })
     })
   })