+++ CHANGELOG.md
@@ -29,11 +29,18 @@ The document follows the conventions described in [“Keep a CHANGELOG”](http:
 - fixed a bug in IE11 with `mouseenter` and `mouseleave` -> __TODO!__
 
 
-## UNRELEASED [2.5.1]
+## UNRELEASED [2.5.2]
+
+### Fixed
+- `clear()` does not remove the parser in svg documents anymore
+
+## [2.5.1] - 2017-03-27
 
 ### Fixed
 - fixed `SVG.PathArray.parse` that did not correctly parsed flat arrays
 
+### Changed
+- make svgjs ready to be used on the server
 
 ## [2.5.0](https://github.com/svgdotjs/svg.js/releases/tag/2.5.0) - 2017-03-10